Input field formatted with the added library
